requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
google-generativeai>=0.8.0
pydantic>=2.8.0
requests>=2.32.0
python-dotenv>=1.0.1
typer>=0.12.5
rich>=13.7.1
jinja2>=3.1.4
pytest>=8.3.2
